+++ __tests__/server/utils/onModuleLoad.spec.jsx
@@ -26,11 +26,12 @@ import onModuleLoad, {
 // This named export exists only on the mock
 // eslint-disable-next-line import/named
 import { setStateConfig, getClientStateConfig, getServerStateConfig } from '../../../src/server/utils/stateConfig';
+import { setRedirectAllowList } from '../../../src/server/utils/redirectAllowList';
 import { setCorsOrigins } from '../../../src/server/plugins/conditionallyAllowCors';
 import { extendRestrictedAttributesAllowList, validateSafeRequestRestrictedAttributes } from '../../../src/server/utils/safeRequest';
 import { setConfigureRequestLog } from '../../../src/server/utils/logging/fastifyPlugin';
 import { setCreateSsrFetch } from '../../../src/server/utils/createSsrFetch';
-import { getEventLoopDelayThreshold } from '../../../src/server/utils/createCircuitBreaker';
+import { getEventLoopDelayThreshold, getEventLoopDelayPercentile } from '../../../src/server/utils/createCircuitBreaker';
 import setupDnsCache from '../../../src/server/utils/setupDnsCache';
 import { configurePWA } from '../../../src/server/pwa';
 import { setErrorPage } from '../../../src/server/plugins/reactHtml/staticErrorPage';
@@ -39,6 +40,9 @@ jest.mock('../../../src/server/utils/stateConfig', () => ({
   setStateConfig: jest.fn(),
   getClientStateConfig: jest.fn(),
   getServerStateConfig: jest.fn(() => ({ rootModuleName: 'root-module' })),
+}));
+jest.mock('../../../src/server/utils/redirectAllowList', () => ({
+  setRedirectAllowList: jest.fn(),
 }));
 jest.mock('@americanexpress/env-config-utils');
 jest.mock('../../../src/server/utils/readJsonFile', () => () => ({ buildVersion: '4.43.0-0-38f0178d' }));
@@ -179,6 +183,31 @@ describe('onModuleLoad', () => {
       moduleName: 'some-root',
     });
     expect(setStateConfig).toHaveBeenCalledWith(provideStateConfig);
+  });
+  it('calls setRedirectAllowList if setRedirectAllowList is supplied', () => {
+    onModuleLoad({
+      module: {
+        [CONFIGURATION_KEY]: {
+          redirectAllowList: ['https://americanexpress.com'],
+          csp,
+        },
+        [META_DATA_KEY]: { version: '1.0.5' },
+      },
+      moduleName: 'some-root',
+    });
+    expect(setRedirectAllowList).toHaveBeenCalledWith(['https://americanexpress.com']);
+  });
+  it('calls setRedirectAllowList with empty array if setRedirectAllowList is NOT supplied', () => {
+    onModuleLoad({
+      module: {
+        [CONFIGURATION_KEY]: {
+          csp,
+        },
+        [META_DATA_KEY]: { version: '1.0.5' },
+      },
+      moduleName: 'some-root',
+    });
+    expect(setRedirectAllowList).toHaveBeenCalledWith([]);
   });
 
   it('does not throw if the root module provides the expected versions of required externals', () => {
@@ -374,7 +403,7 @@ describe('onModuleLoad', () => {
     expect(setErrorPage).toHaveBeenCalledWith(errorPageUrl);
   });
 
-  it('sets the event loop lag threshold from the root module', () => {
+  it('sets the event loop delay threshold from the root module', () => {
     const eventLoopDelayThreshold = 50;
     expect(getEventLoopDelayThreshold()).not.toBe(eventLoopDelayThreshold);
     onModuleLoad({
@@ -388,6 +417,22 @@ describe('onModuleLoad', () => {
       moduleName: 'some-root',
     });
     expect(getEventLoopDelayThreshold()).toBe(eventLoopDelayThreshold);
+  });
+
+  it('sets the event loop delay percentile from the root module', () => {
+    const eventLoopDelayPercentile = 95;
+    expect(getEventLoopDelayPercentile()).not.toBe(eventLoopDelayPercentile);
+    onModuleLoad({
+      module: {
+        [CONFIGURATION_KEY]: {
+          csp,
+          eventLoopDelayPercentile,
+        },
+        [META_DATA_KEY]: { version: '1.0.14' },
+      },
+      moduleName: 'some-root',
+    });
+    expect(getEventLoopDelayPercentile()).toBe(eventLoopDelayPercentile);
   });
 
   it('logs when the root module is loaded', () => {